requirements.txt
twitchio
ollama
python-dotenv